--- app/src/main/res/layout/fragment_preview.xml
@@ -1,128 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<ScrollView xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:fillViewport="true"
-    android:background="?android:attr/colorBackground">
-
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:gravity="top"
-        android:orientation="vertical"
-        android:paddingHorizontal="24dp"
-        android:paddingTop="8dp">
-
-        <FrameLayout
-            android:layout_width="match_parent"
-            android:layout_height="200dp">
-
-            <com.google.android.material.imageview.ShapeableImageView
-                android:id="@+id/image_preview"
-                android:layout_width="match_parent"
-                android:layout_height="match_parent"
-                android:contentDescription="@string/preview"
-                android:scaleType="fitCenter"
-                app:contentPadding="8dp"
-                app:strokeWidth="2dp" />
-
-            <TextView
-                android:id="@+id/image_overlay_text"
-                android:layout_width="match_parent"
-                android:layout_height="match_parent"
-                android:gravity="center"
-                android:textAppearance="@style/TextAppearance.AppCompat.Large"
-                android:visibility="gone"
-                android:background="@android:color/transparent" />
-
-        </FrameLayout>
-
-        <EditText
-            android:id="@+id/file_name"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_marginTop="8dp"
-            android:autofillHints="File Name"
-            android:ellipsize="end"
-            android:gravity="center"
-            android:hint="@string/file_name"
-            android:inputType="textNoSuggestions"
-            android:maxLines="1"
-            android:singleLine="true"
-            android:textAlignment="center"
-            android:textAppearance="@style/TextAppearance.AppCompat.Large"
-            tools:text="test_file.jpg" />
-
-        <Button
-            android:id="@+id/upload_button"
-            android:layout_width="match_parent"
-            android:layout_height="56dp"
-            android:layout_marginTop="8dp"
-            android:paddingEnd="36dp"
-            android:paddingStart="36dp"
-            android:text="@string/upload"
-            android:textAppearance="@style/TextAppearance.AppCompat.Large"
-            app:icon="@drawable/baseline_file_upload_24"
-            app:iconGravity="end"
-            app:iconPadding="-36dp"
-            app:iconSize="24dp" />
-
-        <Button
-            android:id="@+id/options_button"
-            android:layout_width="match_parent"
-            android:layout_height="48dp"
-            android:layout_marginTop="8dp"
-            android:paddingEnd="36dp"
-            android:paddingStart="36dp"
-            android:text="@string/options"
-            android:textAppearance="@style/TextAppearance.AppCompat.Medium"
-            app:icon="@drawable/baseline_settings_24"
-            app:iconGravity="end"
-            app:iconPadding="-36dp"
-            app:iconSize="24dp" />
-
-        <LinearLayout
-            android:layout_width="match_parent"
-            android:layout_height="48dp"
-            android:layout_marginTop="8dp"
-            android:orientation="horizontal">
-
-            <Space
-                android:layout_width="0dp"
-                android:layout_height="match_parent"
-                android:layout_weight="1" />
-
-            <Button
-                android:id="@+id/share_button"
-                android:layout_width="140dp"
-                android:layout_height="match_parent"
-                android:text="@string/share"
-                app:icon="@drawable/baseline_share_24"
-                app:iconGravity="textEnd"
-                app:iconSize="24dp" />
-
-            <Space
-                android:layout_width="0dp"
-                android:layout_height="match_parent"
-                android:layout_weight="1" />
-
-            <Button
-                android:id="@+id/open_button"
-                android:layout_width="140dp"
-                android:layout_height="match_parent"
-                android:text="@string/open"
-                app:icon="@drawable/baseline_open_in_new_24"
-                app:iconGravity="textEnd"
-                app:iconSize="24dp" />
-
-            <Space
-                android:layout_width="0dp"
-                android:layout_height="match_parent"
-                android:layout_weight="1" />
-
-        </LinearLayout>
-
-    </LinearLayout>
-</ScrollView>
